requirements.txt
gunicorn
flask
keras
tensorflow==2.*
pillow
numpy
protobuf==3.*